+++ src/main/java/myjournal/TaskList.java
@@ -1,9 +1,10 @@
 package myjournal;
 
+import myjournal.task.Deadline;
+import myjournal.task.Event;
+import myjournal.task.Task;
+
 import java.util.ArrayList;
-import myjournal.Task;
-import myjournal.Event;
-import myjournal.Deadline;
 
 public class TaskList {
     private ArrayList<Task> tasks;
