+++ src/main/java/myjournal/task/Event.java
@@ -1,7 +1,6 @@
-package myjournal;
+package myjournal.task;
 
-import java.util.Scanner;
-import myjournal.Task;
+import myjournal.task.Task;
 
 /**
  * A class to create tasks with the type of event.
