+++ src/main/java/myjournal/task/Task.java
@@ -1,4 +1,4 @@
-package myjournal;
+package myjournal.task;
 
 /**
  * An abstract class for tasks.
